Refactor lambda plugin
