+++ data-prepper-plugins/lambda/src/main/java/org/opensearch/dataprepper/plugins/sink/lambda/LambdaSinkConfig.java
@@ -1,0 +1,80 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package org.opensearch.dataprepper.plugins.sink.lambda;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import org.opensearch.dataprepper.model.configuration.PluginModel;
+import org.opensearch.dataprepper.plugins.common.lambda.config.AwsAuthenticationOptions;
+import org.opensearch.dataprepper.plugins.common.lambda.config.BatchOptions;
+
+import java.util.Map;
+import java.util.Objects;
+
+public class LambdaSinkConfig {
+
+    private static final int DEFAULT_CONNECTION_RETRIES = 3;
+
+    public static final String STS_REGION = "region";
+
+    public static final String STS_ROLE_ARN = "sts_role_arn";
+
+    @JsonProperty("aws")
+    @NotNull
+    @Valid
+    private AwsAuthenticationOptions awsAuthenticationOptions;
+
+    @JsonProperty("function_name")
+    @NotEmpty
+    @NotNull
+    @Size(min = 3, max = 500, message = "function name length should be at least 3 characters")
+    private String functionName;
+
+    @JsonProperty("max_retries")
+    private int maxConnectionRetries = DEFAULT_CONNECTION_RETRIES;
+
+    @JsonProperty("dlq")
+    private PluginModel dlq;
+
+    @JsonProperty("batch")
+    private BatchOptions batchOptions;
+
+    public AwsAuthenticationOptions getAwsAuthenticationOptions() {
+        return awsAuthenticationOptions;
+    }
+
+    public BatchOptions getBatchOptions(){return batchOptions;}
+
+    public String getFunctionName() {
+        return functionName;
+    }
+
+    public int getMaxConnectionRetries() {
+        return maxConnectionRetries;
+    }
+
+    public PluginModel getDlq() {
+        return dlq;
+    }
+
+    public String getDlqStsRoleARN(){
+        return Objects.nonNull(getDlqPluginSetting().get(STS_ROLE_ARN)) ?
+                String.valueOf(getDlqPluginSetting().get(STS_ROLE_ARN)) :
+                awsAuthenticationOptions.getAwsStsRoleArn();
+    }
+
+    public String getDlqStsRegion(){
+        return Objects.nonNull(getDlqPluginSetting().get(STS_REGION)) ?
+                String.valueOf(getDlqPluginSetting().get(STS_REGION)) :
+                awsAuthenticationOptions.getAwsRegion().toString();
+    }
+
+    public  Map<String, Object> getDlqPluginSetting(){
+        return dlq != null ? dlq.getPluginSettings() : Map.of();
+    }
+}
