+++ data-prepper-plugins/lambda/src/test/resources/mockito-extensions/org.mockito.plugins.MockMaker
