+++ settings.gradle
@@ -175,4 +175,4 @@ include 'data-prepper-plugins:mongodb'
 include 'data-prepper-plugins:rds-source'
 include 'data-prepper-plugins:http-source-common'
 include 'data-prepper-plugins:http-common'
-include 'data-prepper-plugins:lambda-sink'
+include 'data-prepper-plugins:lambda'
